refactor: Move unit tests from integration folder to proper location
- Move backspace cursor synchronization tests to src/repl/view_models/tests.rs
- Remove misplaced test_backspace_display_cursor.rs from tests/ folder
- Update view_models/mod.rs to include tests module
- Replace outdated unit tests with current, focused backspace tests
- All 7 unit tests now pass in proper location
- Fix clippy module_inception warning by renaming module

Follows Rust best practices:
- Unit tests live near the code they test
- Integration tests stay in tests/ folder for cross-module testing
- Cleaner separation of test concerns
