+++ src/repl/view_models/tests.rs
@@ -1,1605 +1,219 @@
-//! # View Model Integration Tests
+//! # View Model Unit Tests
 //!
-//! Tests for the complete view model functionality including ex commands.
+//! Tests for the view model functionality with focus on cursor synchronization and backspace operations.
 
 #[cfg(test)]
-mod integration_tests {
+mod view_model_tests {
     use crate::repl::{
-        commands::CommandEvent,
-        events::{EditorMode, LogicalPosition, Pane, ViewEvent},
+        events::{EditorMode, LogicalPosition},
         view_models::ViewModel,
     };
 
-    #[test]
-    fn test_ex_command_mode_entry_and_exit() {
-        let mut vm = ViewModel::new();
-
-        // Start in normal mode
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-        assert_eq!(vm.get_ex_command_buffer(), "");
-
-        // Enter command mode
-        vm.change_mode(EditorMode::Command).unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Command);
-
-        // Add some characters
-        vm.add_ex_command_char('q').unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), "q");
-
-        // Execute command (should quit and return to normal mode)
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-        assert_eq!(vm.get_ex_command_buffer(), ""); // Buffer should be cleared
-        assert_eq!(events.len(), 1);
-        assert!(matches!(events[0], CommandEvent::QuitRequested));
-    }
+    // === Backspace Display Cursor Synchronization Tests ===
+    // Tests for backspace display cursor synchronization
+    //
+    // These tests verify that the display cursor is properly synchronized
+    // with the logical cursor after backspace operations, preventing the
+    // "black out" issue when the Response pane is rendered.
 
     #[test]
-    fn test_ex_command_quit_variations() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Test "q" command
-        vm.add_ex_command_char('q').unwrap();
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 1);
-        assert!(matches!(events[0], CommandEvent::QuitRequested));
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-
-        // Test "q!" command
-        vm.change_mode(EditorMode::Command).unwrap();
-        vm.add_ex_command_char('q').unwrap();
-        vm.add_ex_command_char('!').unwrap();
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 1);
-        assert!(matches!(events[0], CommandEvent::QuitRequested));
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-    }
-
-    #[test]
-    fn test_ex_command_wrap_commands() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Test "set wrap" command
-        for ch in "set wrap".chars() {
-            vm.add_ex_command_char(ch).unwrap();
-        }
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 0); // No command events, just internal state change
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-
-        // Test "set nowrap" command
-        vm.change_mode(EditorMode::Command).unwrap();
-        for ch in "set nowrap".chars() {
-            vm.add_ex_command_char(ch).unwrap();
-        }
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 0);
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-    }
-
-    #[test]
-    fn test_ex_command_empty_command() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Execute empty command (just pressing Enter without typing)
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 0);
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-        assert_eq!(vm.get_ex_command_buffer(), "");
-    }
-
-    #[test]
-    fn test_ex_command_unknown_command() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Type unknown command
-        for ch in "unknown".chars() {
-            vm.add_ex_command_char(ch).unwrap();
-        }
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 0); // No events for unknown commands
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-        assert_eq!(vm.get_ex_command_buffer(), "");
-    }
-
-    #[test]
-    fn test_ex_command_backspace() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Type some characters
-        vm.add_ex_command_char('q').unwrap();
-        vm.add_ex_command_char('u').unwrap();
-        vm.add_ex_command_char('i').unwrap();
-        vm.add_ex_command_char('t').unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), "quit");
-
-        // Backspace to remove last character
-        vm.backspace_ex_command().unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), "qui");
-
-        // Backspace again
-        vm.backspace_ex_command().unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), "qu");
-
-        // Continue backspacing until empty
-        vm.backspace_ex_command().unwrap();
-        vm.backspace_ex_command().unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), "");
-
-        // Backspace on empty buffer should not panic
-        vm.backspace_ex_command().unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), "");
-    }
-
-    #[test]
-    fn test_ex_command_whitespace_handling() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Test command with leading/trailing whitespace
-        vm.add_ex_command_char(' ').unwrap();
-        vm.add_ex_command_char('q').unwrap();
-        vm.add_ex_command_char(' ').unwrap();
-        assert_eq!(vm.get_ex_command_buffer(), " q ");
-
-        // Execute - should still work due to trim()
-        let events = vm.execute_ex_command().unwrap();
-        assert_eq!(events.len(), 1);
-        assert!(matches!(events[0], CommandEvent::QuitRequested));
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-    }
-
-    #[test]
-    fn test_delete_char_before_cursor_in_line() {
+    fn test_backspace_updates_display_cursor_correctly() {
         let mut vm = ViewModel::new();
         vm.change_mode(EditorMode::Insert).unwrap();
 
-        // Insert some text
-        vm.insert_text("hello").unwrap();
+        // Insert "GET" in the request pane
+        vm.insert_text("GET").unwrap();
 
-        // Delete a character (should delete 'o')
+        // Verify initial positions
+        let logical_cursor = vm.get_cursor_position();
+        assert_eq!(logical_cursor.line, 0);
+        assert_eq!(logical_cursor.column, 3); // After "GET"
+
+        let display_cursor = vm.get_display_cursor_position();
+        assert_eq!(display_cursor.0, 0); // Display line
+        assert_eq!(display_cursor.1, 3); // Display column
+
+        // Perform backspace
         vm.delete_char_before_cursor().unwrap();
 
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.column, 4); // Should be at position 4 (after "hell")
+        // Verify logical cursor moved correctly
+        let logical_cursor_after = vm.get_cursor_position();
+        assert_eq!(logical_cursor_after.line, 0);
+        assert_eq!(logical_cursor_after.column, 2); // After "GE"
+
+        // Verify display cursor is synchronized
+        let display_cursor_after = vm.get_display_cursor_position();
+        assert_eq!(display_cursor_after.0, 0); // Display line
+        assert_eq!(display_cursor_after.1, 2); // Display column - MUST match logical
+
+        // Verify the text content
+        assert_eq!(vm.get_request_text(), "GE");
     }
 
     #[test]
-    fn test_delete_char_before_cursor_at_line_start() {
+    fn test_backspace_at_line_join_updates_display_cursor() {
         let mut vm = ViewModel::new();
         vm.change_mode(EditorMode::Insert).unwrap();
 
-        // Insert simple text and test basic backspace
-        vm.insert_text("hello").unwrap();
-
-        // Move cursor to position 1
-        vm.set_cursor_position(LogicalPosition::new(0, 1)).unwrap();
-
-        // Delete character before cursor (should delete 'h')
-        vm.delete_char_before_cursor().unwrap();
-
-        // Should be at position 0 now
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 0);
-    }
-
-    #[test]
-    fn test_delete_char_before_cursor_only_in_insert_mode() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text in insert mode
-        vm.insert_text("test").unwrap();
-        // Switch to normal mode
-        vm.change_mode(EditorMode::Normal).unwrap();
-
-        // Try to delete in normal mode (should do nothing)
-        vm.delete_char_before_cursor().unwrap();
-
-        // Text should be unchanged
-        assert_eq!(vm.get_request_text(), "test");
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text and move cursor back
-        vm.insert_text("hello").unwrap();
-        let cursor = vm.get_cursor_position();
-        vm.set_cursor_position(LogicalPosition::new(cursor.line, 2))
-            .unwrap();
-
-        // Delete character after cursor (should delete 'l')
-        vm.delete_char_after_cursor().unwrap();
-
-        // Cursor should stay at same position but text should be modified
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.column, 2);
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_at_line_end() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text and position at end
-        vm.insert_text("hello").unwrap();
-
-        // Try to delete at end of line (should do nothing)
-        vm.delete_char_after_cursor().unwrap();
-
-        // Text should be unchanged
-        assert_eq!(vm.get_request_text(), "hello");
-    }
-
-    #[test]
-    fn test_cursor_moves_correctly_after_backspace() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text "hello" (cursor should be at position 5)
-        vm.insert_text("hello").unwrap();
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.column, 5);
-
-        // Backspace once (should delete 'o' and move cursor to position 4)
-        vm.delete_char_before_cursor().unwrap();
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.column, 4);
-
-        // Backspace again (should delete 'l' and move cursor to position 3)
-        vm.delete_char_before_cursor().unwrap();
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.column, 3);
-
-        // Verify the text is now "hel"
-        assert_eq!(vm.get_request_text(), "hel");
-    }
-
-    #[test]
-    fn test_delete_char_before_cursor_on_blank_line() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with a blank line: "GET /api/users\n\n{"name": "John"}"
-        vm.insert_text("GET /api/users\n\n{\"name\": \"John\"}")
-            .unwrap();
-
-        // Position cursor on the blank line (line 1, column 0)
-        vm.set_cursor_position(LogicalPosition::new(1, 0)).unwrap();
-
-        // Backspace should delete the blank line and move cursor to end of previous line
-        vm.delete_char_before_cursor().unwrap();
-
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 14); // Should be at end of "GET /api/users"
-
-        // Verify the blank line was deleted
-        assert_eq!(
-            vm.get_request_text(),
-            "GET /api/users\n{\"name\": \"John\"}"
-        );
-    }
-
-    #[test]
-    fn test_delete_char_before_cursor_on_consecutive_blank_lines() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with consecutive blank lines: "GET /api/users\n\n\n{"name": "John"}"
-        vm.insert_text("GET /api/users\n\n\n{\"name\": \"John\"}")
-            .unwrap();
-
-        // Position cursor on the second blank line (line 2, column 0)
-        vm.set_cursor_position(LogicalPosition::new(2, 0)).unwrap();
-
-        // Backspace should delete only the current blank line
-        vm.delete_char_before_cursor().unwrap();
-
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 1);
-        assert_eq!(cursor.column, 0); // Should be at the beginning of the first blank line
-
-        // Verify only one blank line was deleted
-        assert_eq!(
-            vm.get_request_text(),
-            "GET /api/users\n\n{\"name\": \"John\"}"
-        );
-    }
-
-    #[test]
-    fn test_delete_char_before_cursor_blank_line_moves_to_end_of_previous() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text where previous line has content
-        vm.insert_text("Hello World\n").unwrap();
-
-        // Position cursor on the blank line (line 1, column 0)
-        vm.set_cursor_position(LogicalPosition::new(1, 0)).unwrap();
-
-        // Backspace should delete the blank line and move cursor to end of "Hello World"
-        vm.delete_char_before_cursor().unwrap();
-
-        let cursor = vm.get_cursor_position();
-
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 11); // Should be at end of "Hello World"
-
-        // Verify the blank line was deleted
-        assert_eq!(vm.get_request_text(), "Hello World");
-    }
-
-    #[test]
-    fn test_delete_char_before_cursor_non_blank_line_joins_normally() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with non-blank second line
-        vm.insert_text("Hello\nWorld").unwrap();
+        // Insert multiline text
+        vm.insert_text("GET /api\nHost: example.com").unwrap();
 
         // Position cursor at start of second line
         vm.set_cursor_position(LogicalPosition::new(1, 0)).unwrap();
 
-        // Backspace should join the lines (existing behavior)
+        // Verify initial display cursor
+        let display_cursor = vm.get_display_cursor_position();
+        assert_eq!(display_cursor.0, 1); // Second display line
+        assert_eq!(display_cursor.1, 0); // Start of line
+
+        // Perform backspace (should join lines)
         vm.delete_char_before_cursor().unwrap();
 
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 5); // Should be after "Hello"
+        // Verify logical cursor is at end of joined first line
+        let logical_cursor_after = vm.get_cursor_position();
+        assert_eq!(logical_cursor_after.line, 0);
+        assert_eq!(logical_cursor_after.column, 8); // After "GET /api"
 
-        // Verify lines were joined
-        assert_eq!(vm.get_request_text(), "HelloWorld");
+        // Verify display cursor is synchronized
+        let display_cursor_after = vm.get_display_cursor_position();
+        assert_eq!(display_cursor_after.0, 0); // First display line
+        assert_eq!(display_cursor_after.1, 8); // After "GET /api"
+
+        // Verify the text was joined correctly
+        assert_eq!(vm.get_request_text(), "GET /apiHost: example.com");
     }
 
     #[test]
-    fn test_pane_switching() {
-        let mut vm = ViewModel::new();
-
-        // Should start in request pane
-        assert_eq!(vm.get_current_pane(), crate::repl::events::Pane::Request);
-
-        // Add a response so response pane is available
-        vm.set_response(200, "test response".to_string());
-
-        // Switch to response pane
-        vm.switch_pane(crate::repl::events::Pane::Response).unwrap();
-        assert_eq!(vm.get_current_pane(), crate::repl::events::Pane::Response);
-
-        // Switch back to request pane
-        vm.switch_pane(crate::repl::events::Pane::Request).unwrap();
-        assert_eq!(vm.get_current_pane(), crate::repl::events::Pane::Request);
-    }
-
-    #[test]
-    fn test_navigation_in_response_pane() {
-        let mut vm = ViewModel::new();
-
-        // Set up a response with multiple lines
-        vm.set_response(200, "line 1\nline 2\nline 3".to_string());
-
-        // Switch to response pane and normal mode
-        vm.switch_pane(crate::repl::events::Pane::Response).unwrap();
-        vm.change_mode(EditorMode::Normal).unwrap();
-
-        assert_eq!(vm.get_current_pane(), crate::repl::events::Pane::Response);
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-
-        // Movement should work in response pane
-        let initial_cursor = vm.get_cursor_position();
-        assert_eq!(initial_cursor.line, 0);
-        assert_eq!(initial_cursor.column, 0);
-
-        // Try moving down
-        vm.move_cursor_down().unwrap();
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 1); // Should move to line 1
-    }
-
-    #[test]
-    fn test_command_flow_integration() {
-        use crate::repl::commands::{CommandContext, CommandRegistry, ViewModelSnapshot};
-        use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
-
-        let mut vm = ViewModel::new();
-        let registry = CommandRegistry::new();
-
-        // Set up a response
-        vm.set_response(200, "line 1\nline 2\nline 3".to_string());
-
-        // Create context from view model state
-        let context = CommandContext::new(ViewModelSnapshot::from_view_model(&vm));
-
-        // Test Tab key (should generate PaneSwitchRequested)
-        let tab_event = KeyEvent::new(KeyCode::Tab, KeyModifiers::empty());
-        let events = registry.process_event(tab_event, &context).unwrap();
-        assert!(!events.is_empty(), "Tab should generate events");
-
-        // Test j key in normal mode (should generate cursor move)
-        vm.change_mode(EditorMode::Normal).unwrap();
-        let context = CommandContext::new(ViewModelSnapshot::from_view_model(&vm));
-        let j_event = KeyEvent::new(KeyCode::Char('j'), KeyModifiers::empty());
-        let events = registry.process_event(j_event, &context).unwrap();
-        assert!(
-            !events.is_empty(),
-            "j key should generate cursor move events"
-        );
-    }
-
-    #[test]
-    fn test_line_navigation_with_valid_line_number() {
+    fn test_multiple_backspaces_maintain_display_cursor_sync() {
         let mut vm = ViewModel::new();
         vm.change_mode(EditorMode::Insert).unwrap();
 
-        // Add multi-line content to test navigation
-        vm.insert_text("line 1\nline 2\nline 3\nline 4").unwrap();
-        vm.change_mode(EditorMode::Command).unwrap();
+        // Insert text
+        vm.insert_text("HELLO").unwrap();
 
-        // Navigate to line 2 using :2 command
-        vm.add_ex_command_char('2').unwrap();
-        let events = vm.execute_ex_command().unwrap();
+        // Perform multiple backspaces
+        for i in 0..3 {
+            vm.delete_char_before_cursor().unwrap();
 
-        // Should get a cursor move event
-        assert_eq!(events.len(), 1);
-        if let CommandEvent::CursorMoveRequested { direction, amount } = &events[0] {
-            assert_eq!(*amount, 1);
-            if let crate::repl::commands::MovementDirection::LineNumber(line_num) = direction {
-                assert_eq!(*line_num, 2);
-            } else {
-                panic!("Expected LineNumber movement direction, got {direction:?}");
-            }
-        } else {
-            panic!("Expected CursorMoveRequested event, got {:?}", events[0]);
+            // After each backspace, verify display cursor matches logical cursor
+            let logical_cursor = vm.get_cursor_position();
+            let display_cursor = vm.get_display_cursor_position();
+
+            assert_eq!(logical_cursor.line, 0);
+            assert_eq!(logical_cursor.column, 5 - (i + 1)); // 4, 3, 2
+            assert_eq!(display_cursor.0, 0);
+            assert_eq!(display_cursor.1, 5 - (i + 1)); // Must match logical column
         }
 
-        // Should be back in normal mode
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-        assert_eq!(vm.get_ex_command_buffer(), ""); // Buffer should be cleared
+        // Verify final text
+        assert_eq!(vm.get_request_text(), "HE");
     }
 
     #[test]
-    fn test_line_navigation_with_line_number_one() {
+    fn test_backspace_with_wrapped_lines_updates_display_cursor() {
         let mut vm = ViewModel::new();
+
+        // Set up narrow terminal to force wrapping
+        vm.update_terminal_size(20, 10);
+        vm.set_wrap_enabled(true).unwrap();
+
         vm.change_mode(EditorMode::Insert).unwrap();
 
-        // Add multi-line content
-        vm.insert_text("line 1\nline 2\nline 3").unwrap();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Navigate to line 1 using :1 command
-        vm.add_ex_command_char('1').unwrap();
-        let events = vm.execute_ex_command().unwrap();
-
-        // Should get a cursor move event
-        assert_eq!(events.len(), 1);
-        if let CommandEvent::CursorMoveRequested {
-            direction,
-            amount: _,
-        } = &events[0]
-        {
-            if let crate::repl::commands::MovementDirection::LineNumber(line_num) = direction {
-                assert_eq!(*line_num, 1);
-            } else {
-                panic!("Expected LineNumber movement direction");
-            }
-        } else {
-            panic!("Expected CursorMoveRequested event");
-        }
-
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-    }
-
-    #[test]
-    fn test_line_navigation_with_zero_line_number_should_be_ignored() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Try to navigate to line 0 - should be ignored
-        vm.add_ex_command_char('0').unwrap();
-        let events = vm.execute_ex_command().unwrap();
-
-        // Should not generate any cursor move events (line 0 is invalid)
-        assert_eq!(events.len(), 0);
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-    }
-
-    #[test]
-    fn test_line_navigation_with_large_line_number() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Add only 2 lines
-        vm.insert_text("line 1\nline 2").unwrap();
-        vm.change_mode(EditorMode::Command).unwrap();
-
-        // Try to navigate to line 100 - should work (cursor will be clamped)
-        vm.add_ex_command_char('1').unwrap();
-        vm.add_ex_command_char('0').unwrap();
-        vm.add_ex_command_char('0').unwrap();
-        let events = vm.execute_ex_command().unwrap();
-
-        // Should get a cursor move event for line 100
-        assert_eq!(events.len(), 1);
-        if let CommandEvent::CursorMoveRequested {
-            direction,
-            amount: _,
-        } = &events[0]
-        {
-            if let crate::repl::commands::MovementDirection::LineNumber(line_num) = direction {
-                assert_eq!(*line_num, 100);
-            } else {
-                panic!("Expected LineNumber movement direction");
-            }
-        } else {
-            panic!("Expected CursorMoveRequested event");
-        }
-
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-    }
-
-    #[test]
-    fn test_move_cursor_to_line_method_with_valid_line() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Add content with multiple lines
-        vm.insert_text("line 1\nline 2\nline 3\nline 4\nline 5")
+        // Insert a long line that will wrap
+        vm.insert_text("This is a very long line that will wrap")
             .unwrap();
 
-        // Move to line 3
-        vm.move_cursor_to_line(3).unwrap();
+        // Get cursor position (should be at end of text)
+        let logical_cursor = vm.get_cursor_position();
+        assert_eq!(logical_cursor.line, 0);
+        assert_eq!(logical_cursor.column, 39); // Length of text "This is a very long line that will wrap" = 39
 
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 2); // 0-based indexing, so line 3 = index 2
-        assert_eq!(cursor.column, 0); // Should be at beginning of line
+        // Perform backspace
+        vm.delete_char_before_cursor().unwrap();
+
+        // Verify logical cursor moved back
+        let logical_cursor_after = vm.get_cursor_position();
+        assert_eq!(logical_cursor_after.line, 0);
+        assert_eq!(logical_cursor_after.column, 38);
+
+        // Display cursor should be on a wrapped line but properly synchronized
+        let display_cursor_after = vm.get_display_cursor_position();
+        // The exact display line/column depends on wrapping implementation
+        // but the key is that it should not be stale/incorrect
+
+        // Perform another backspace to ensure continued sync
+        vm.delete_char_before_cursor().unwrap();
+
+        let logical_cursor_after2 = vm.get_cursor_position();
+        let display_cursor_after2 = vm.get_display_cursor_position();
+
+        assert_eq!(logical_cursor_after2.column, 37);
+        // Display cursor should have moved as well (exact position depends on wrap points)
+        assert!(display_cursor_after2 != display_cursor_after);
     }
 
     #[test]
-    fn test_move_cursor_to_line_method_with_line_beyond_buffer() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Add only 2 lines
-        vm.insert_text("line 1\nline 2").unwrap();
-
-        // Try to move to line 10 (beyond buffer)
-        vm.move_cursor_to_line(10).unwrap();
-
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 1); // Should be clamped to last line (0-based)
-        assert_eq!(cursor.column, 0);
-    }
-
-    #[test]
-    fn test_move_cursor_to_line_method_with_zero_should_do_nothing() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        vm.insert_text("line 1\nline 2").unwrap();
-
-        // Get initial cursor position
-        let initial_cursor = vm.get_cursor_position();
-
-        // Try to move to line 0 (invalid)
-        vm.move_cursor_to_line(0).unwrap();
-
-        // Cursor should not have moved
-        let final_cursor = vm.get_cursor_position();
-        assert_eq!(initial_cursor, final_cursor);
-    }
-
-    #[test]
-    fn test_show_profile_command() {
+    fn test_backspace_preserves_display_cursor_sync_with_response_pane() {
         let mut vm = ViewModel::new();
 
-        // Set custom profile info
-        vm.set_profile_info(
-            "test-profile".to_string(),
-            "/custom/path/profile".to_string(),
+        // Set up response content to ensure response pane exists
+        vm.set_response(
+            200,
+            "HTTP/1.1 200 OK\nContent-Type: application/json\n\n{\"status\": \"ok\"}".to_string(),
         );
 
-        // Enter command mode and execute :show profile
-        vm.change_mode(EditorMode::Command).unwrap();
-        vm.add_ex_command_char('s').unwrap();
-        vm.add_ex_command_char('h').unwrap();
-        vm.add_ex_command_char('o').unwrap();
-        vm.add_ex_command_char('w').unwrap();
-        vm.add_ex_command_char(' ').unwrap();
-        vm.add_ex_command_char('p').unwrap();
-        vm.add_ex_command_char('r').unwrap();
-        vm.add_ex_command_char('o').unwrap();
-        vm.add_ex_command_char('f').unwrap();
-        vm.add_ex_command_char('i').unwrap();
-        vm.add_ex_command_char('l').unwrap();
-        vm.add_ex_command_char('e').unwrap();
+        // Make sure we're in request pane (we're already there by default)
+        vm.change_mode(EditorMode::Insert).unwrap();
 
-        let events = vm.execute_ex_command().unwrap();
+        // Insert text in request pane
+        vm.insert_text("GET /api/test").unwrap();
 
-        // Should get a ShowProfileRequested event
-        assert_eq!(events.len(), 1);
-        assert!(matches!(events[0], CommandEvent::ShowProfileRequested));
+        // Verify we can get display cursor position
+        let display_cursor_before = vm.get_display_cursor_position();
+        assert_eq!(display_cursor_before.1, 13); // After "GET /api/test"
 
-        // Should be back in normal mode
-        assert_eq!(vm.get_mode(), EditorMode::Normal);
-        assert_eq!(vm.get_ex_command_buffer(), ""); // Buffer should be cleared
+        // Perform backspace
+        vm.delete_char_before_cursor().unwrap();
+
+        // Verify display cursor updated correctly
+        let display_cursor_after = vm.get_display_cursor_position();
+        assert_eq!(display_cursor_after.1, 12); // After "GET /api/tes"
+
+        // The key test: rendering should not "black out" the response pane
+        // This would happen if display cursor was out of sync
+        assert_eq!(vm.get_request_text(), "GET /api/tes");
     }
 
     #[test]
-    fn test_profile_info_getters_and_setters() {
+    fn test_basic_text_insertion() {
         let mut vm = ViewModel::new();
+        vm.change_mode(EditorMode::Insert).unwrap();
 
-        // Check default values
-        assert_eq!(vm.get_profile_name(), "default");
-        assert_eq!(vm.get_profile_path(), "~/.blueline/profile");
+        // Insert text
+        vm.insert_text("Hello, World!").unwrap();
 
-        // Set custom values
-        vm.set_profile_info("custom-profile".to_string(), "/custom/path".to_string());
+        // Verify the text was inserted
+        assert_eq!(vm.get_request_text(), "Hello, World!");
 
-        // Check updated values
-        assert_eq!(vm.get_profile_name(), "custom-profile");
-        assert_eq!(vm.get_profile_path(), "/custom/path");
+        // Verify cursor position
+        let cursor = vm.get_cursor_position();
+        assert_eq!(cursor.line, 0);
+        assert_eq!(cursor.column, 13); // After "Hello, World!"
     }
 
     #[test]
-    fn test_status_message_functionality() {
-        let mut vm = ViewModel::new();
-
-        // Initially no status message
-        assert_eq!(vm.get_status_message(), None);
-
-        // Set a status message
-        vm.set_status_message("Test message");
-        assert_eq!(vm.get_status_message(), Some("Test message"));
-
-        // Set another message
-        vm.set_status_message("Another message".to_string());
-        assert_eq!(vm.get_status_message(), Some("Another message"));
-
-        // Clear the message
-        vm.clear_status_message();
-        assert_eq!(vm.get_status_message(), None);
-    }
-
-    // Visual mode tests
-    #[test]
-    fn test_visual_mode_entry_and_exit() {
+    fn test_mode_transitions() {
         let mut vm = ViewModel::new();
 
         // Start in normal mode
         assert_eq!(vm.get_mode(), EditorMode::Normal);
 
-        // Enter visual mode
-        vm.change_mode(EditorMode::Visual).unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Visual);
+        // Switch to insert mode
+        vm.change_mode(EditorMode::Insert).unwrap();
+        assert_eq!(vm.get_mode(), EditorMode::Insert);
 
-        // Check that visual selection state is initialized
-        let (start, end, pane) = vm.get_visual_selection();
-        assert!(start.is_some());
-        assert!(end.is_some());
-        assert!(pane.is_some());
-        assert_eq!(pane.unwrap(), Pane::Request);
-
-        // Exit visual mode
+        // Switch back to normal mode
         vm.change_mode(EditorMode::Normal).unwrap();
         assert_eq!(vm.get_mode(), EditorMode::Normal);
-
-        // Check that visual selection state is cleared
-        let (start, end, pane) = vm.get_visual_selection();
-        assert!(start.is_none());
-        assert!(end.is_none());
-        assert!(pane.is_none());
-    }
-
-    #[test]
-    fn test_visual_selection_initialization() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Add some text and position cursor
-        vm.insert_text("hello world").unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-
-        // Enter visual mode
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        // Check that selection starts and ends at cursor position
-        let (start, end, pane) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 5)));
-        assert_eq!(end, Some(LogicalPosition::new(0, 5)));
-        assert_eq!(pane, Some(Pane::Request));
-    }
-
-    #[test]
-    fn test_visual_selection_updates_with_cursor_movement() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Add some text
-        vm.insert_text("hello\nworld\ntest").unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 2)).unwrap();
-
-        // Enter visual mode
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        let (start, end, _) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 2)));
-        assert_eq!(end, Some(LogicalPosition::new(0, 2)));
-
-        // Move cursor - selection end should update
-        vm.set_cursor_position(LogicalPosition::new(1, 3)).unwrap();
-
-        let (start, end, _) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 2))); // Start unchanged
-        assert_eq!(end, Some(LogicalPosition::new(1, 3))); // End updated
-    }
-
-    #[test]
-    fn test_is_position_selected_single_line() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world").unwrap();
-
-        // Set up visual selection from column 2 to 6 on line 0
-        vm.change_mode(EditorMode::Visual).unwrap();
-        vm.panes[Pane::Request].visual_selection_start = Some(LogicalPosition::new(0, 2));
-        vm.panes[Pane::Request].visual_selection_end = Some(LogicalPosition::new(0, 6));
-
-        // Test positions within selection
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 2), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 3), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 4), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 5), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 6), Pane::Request));
-
-        // Test positions outside selection
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 1), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 7), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(1, 3), Pane::Request));
-
-        // Test different pane
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 3), Pane::Response));
-    }
-
-    #[test]
-    fn test_is_position_selected_multi_line() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("line1\nline2\nline3").unwrap();
-
-        // Set up visual selection from line 0, col 2 to line 2, col 3
-        vm.change_mode(EditorMode::Visual).unwrap();
-        vm.panes[Pane::Request].visual_selection_start = Some(LogicalPosition::new(0, 2));
-        vm.panes[Pane::Request].visual_selection_end = Some(LogicalPosition::new(2, 3));
-
-        // Test first line (partial)
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 1), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 2), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 3), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 4), Pane::Request));
-
-        // Test middle line (fully selected)
-        assert!(vm.is_position_selected(LogicalPosition::new(1, 0), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(1, 2), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(1, 4), Pane::Request));
-
-        // Test last line (partial)
-        assert!(vm.is_position_selected(LogicalPosition::new(2, 0), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(2, 3), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(2, 4), Pane::Request));
-
-        // Test lines outside selection
-        assert!(!vm.is_position_selected(LogicalPosition::new(3, 0), Pane::Request));
-    }
-
-    #[test]
-    fn test_is_position_selected_reversed_selection() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world").unwrap();
-
-        // Set up visual selection with end before start (reversed)
-        vm.change_mode(EditorMode::Visual).unwrap();
-        vm.panes[Pane::Request].visual_selection_start = Some(LogicalPosition::new(0, 6));
-        vm.panes[Pane::Request].visual_selection_end = Some(LogicalPosition::new(0, 2));
-
-        // Should normalize selection range automatically
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 2), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 3), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 6), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 1), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 7), Pane::Request));
-    }
-
-    #[test]
-    fn test_visual_selection_in_response_pane() {
-        let mut vm = ViewModel::new();
-
-        // Set up response content
-        vm.set_response(200, "response\ntext\nhere".to_string());
-        vm.switch_pane(Pane::Response).unwrap();
-
-        // Enter visual mode in response pane
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        let (_start, _end, pane) = vm.get_visual_selection();
-        assert_eq!(pane, Some(Pane::Response));
-
-        // Move cursor and check selection updates
-        vm.set_cursor_position(LogicalPosition::new(1, 2)).unwrap();
-
-        let (_start, end, pane) = vm.get_visual_selection();
-        assert_eq!(pane, Some(Pane::Response));
-        assert_eq!(end, Some(LogicalPosition::new(1, 2)));
-    }
-
-    #[test]
-    fn test_visual_selection_does_not_update_across_panes() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("request text").unwrap();
-        vm.set_response(200, "response text".to_string());
-
-        // Start visual selection in request pane
-        vm.change_mode(EditorMode::Visual).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-
-        // Switch to response pane - selection should not update
-        vm.switch_pane(Pane::Response).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 3)).unwrap();
-
-        // Selection should still be in request pane and unchanged
-        // With pane-based selection, the current pane (Response) has no selection
-        let (_start, _end, pane) = vm.get_visual_selection();
-        assert_eq!(pane, None); // Response pane has no selection
-
-        // But the request pane should still have its original selection
-        assert_eq!(
-            vm.panes[Pane::Request].visual_selection_end,
-            Some(LogicalPosition::new(0, 5))
-        );
-    }
-
-    #[test]
-    fn test_visual_mode_persists_across_cursor_movements() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello\nworld").unwrap();
-
-        // Position cursor at a specific location before entering visual mode
-        vm.set_cursor_position(LogicalPosition::new(0, 2)).unwrap();
-
-        // Enter visual mode
-        vm.change_mode(EditorMode::Visual).unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Visual);
-
-        let (initial_start, initial_end, _) = vm.get_visual_selection();
-        assert_eq!(initial_start, Some(LogicalPosition::new(0, 2)));
-        assert_eq!(initial_end, Some(LogicalPosition::new(0, 2)));
-
-        // Move cursor multiple times - should stay in visual mode
-        vm.move_cursor_right().unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Visual);
-
-        vm.move_cursor_down().unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Visual);
-
-        vm.move_cursor_right().unwrap();
-        assert_eq!(vm.get_mode(), EditorMode::Visual);
-
-        // Selection should be updated
-        let (start, end, _) = vm.get_visual_selection();
-        assert!(start.is_some());
-        assert!(end.is_some());
-        assert_eq!(start, Some(LogicalPosition::new(0, 2))); // Start should remain unchanged
-        assert_ne!(start, end); // End should be different after movements
-    }
-
-    // Tests for multiple event emission functionality
-    #[test]
-    fn test_emit_view_event_single_event() {
-        let mut vm = ViewModel::new();
-
-        // Emit a single event
-        vm.emit_view_event([ViewEvent::StatusBarUpdateRequired]);
-
-        // Check that it was properly collected
-        let events = vm.collect_pending_view_events();
-        assert_eq!(events.len(), 1);
-        assert!(matches!(events[0], ViewEvent::StatusBarUpdateRequired));
-    }
-
-    #[test]
-    fn test_emit_view_event_multiple_events() {
-        let mut vm = ViewModel::new();
-
-        // Emit multiple events at once
-        vm.emit_view_event([
-            ViewEvent::StatusBarUpdateRequired,
-            ViewEvent::CursorUpdateRequired {
-                pane: Pane::Request,
-            },
-            ViewEvent::PositionIndicatorUpdateRequired,
-        ]);
-
-        // Check that all events were collected
-        let events = vm.collect_pending_view_events();
-        assert_eq!(events.len(), 3);
-        assert!(matches!(events[0], ViewEvent::StatusBarUpdateRequired));
-        assert!(matches!(
-            events[1],
-            ViewEvent::CursorUpdateRequired {
-                pane: Pane::Request
-            }
-        ));
-        assert!(matches!(
-            events[2],
-            ViewEvent::PositionIndicatorUpdateRequired
-        ));
-    }
-
-    #[test]
-    fn test_emit_view_event_vec_of_events() {
-        let mut vm = ViewModel::new();
-
-        // Create a vector of events
-        let event_vec = vec![
-            ViewEvent::PaneRedrawRequired {
-                pane: Pane::Request,
-            },
-            ViewEvent::CursorUpdateRequired {
-                pane: Pane::Response,
-            },
-        ];
-
-        // Emit the vector
-        vm.emit_view_event(event_vec);
-
-        // Check that all events were collected
-        let events = vm.collect_pending_view_events();
-        assert_eq!(events.len(), 2);
-        assert!(matches!(
-            events[0],
-            ViewEvent::PaneRedrawRequired {
-                pane: Pane::Request
-            }
-        ));
-        assert!(matches!(
-            events[1],
-            ViewEvent::CursorUpdateRequired {
-                pane: Pane::Response
-            }
-        ));
-    }
-
-    #[test]
-    fn test_emit_view_event_accumulates_events() {
-        let mut vm = ViewModel::new();
-
-        // Emit some events
-        vm.emit_view_event([ViewEvent::StatusBarUpdateRequired]);
-        vm.emit_view_event([
-            ViewEvent::CursorUpdateRequired {
-                pane: Pane::Request,
-            },
-            ViewEvent::PositionIndicatorUpdateRequired,
-        ]);
-
-        // Check that all events were accumulated
-        let events = vm.collect_pending_view_events();
-        assert_eq!(events.len(), 3);
-
-        // Check that collecting clears the events
-        let empty_events = vm.collect_pending_view_events();
-        assert_eq!(empty_events.len(), 0);
-    }
-
-    // Tests for Delete key empty line functionality (Issue #34)
-    #[test]
-    fn test_delete_char_after_cursor_removes_empty_next_line() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with empty line: "hello\n\nworld"
-        vm.insert_text("hello\n\nworld").unwrap();
-
-        // Position cursor at end of first line (after "hello")
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-
-        // Delete should remove the empty line
-        vm.delete_char_after_cursor().unwrap();
-
-        // Text should now be "hello\nworld" with cursor still at end of first line
-        assert_eq!(vm.get_request_text(), "hello\nworld");
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 5);
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_joins_lines_when_next_line_has_content() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text: "hello\nworld"
-        vm.insert_text("hello\nworld").unwrap();
-
-        // Position cursor at end of first line
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-
-        // Delete should join the lines
-        vm.delete_char_after_cursor().unwrap();
-
-        // Text should now be "helloworld"
-        assert_eq!(vm.get_request_text(), "helloworld");
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 5); // Cursor stays at same position
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_on_empty_line_deletes_current_line() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with empty line in middle: "hello\n\nworld"
-        vm.insert_text("hello\n\nworld").unwrap();
-
-        // Position cursor on the empty line (line 1, column 0)
-        vm.set_cursor_position(LogicalPosition::new(1, 0)).unwrap();
-
-        // Delete should remove the empty line and move cursor to end of previous line
-        vm.delete_char_after_cursor().unwrap();
-
-        // Text should now be "hello\nworld"
-        assert_eq!(vm.get_request_text(), "hello\nworld");
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 5); // Should be at end of "hello"
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_on_empty_first_line_does_nothing() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Start with empty buffer (first line is empty and there's no previous line)
-        // Position cursor at (0, 0)
-        vm.set_cursor_position(LogicalPosition::new(0, 0)).unwrap();
-
-        // Delete should do nothing since there's no previous line to move to
-        vm.delete_char_after_cursor().unwrap();
-
-        // Buffer should remain empty
-        assert_eq!(vm.get_request_text(), "");
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 0);
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_multiple_consecutive_empty_lines() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with multiple empty lines: "hello\n\n\n\nworld"
-        vm.insert_text("hello\n\n\n\nworld").unwrap();
-
-        // Position cursor at end of first line
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-
-        // First delete should remove first empty line
-        vm.delete_char_after_cursor().unwrap();
-        assert_eq!(vm.get_request_text(), "hello\n\n\nworld");
-
-        // Second delete should remove next empty line
-        vm.delete_char_after_cursor().unwrap();
-        assert_eq!(vm.get_request_text(), "hello\n\nworld");
-
-        // Third delete should remove last empty line
-        vm.delete_char_after_cursor().unwrap();
-        assert_eq!(vm.get_request_text(), "hello\nworld");
-
-        // Cursor should remain at end of first line
-        let cursor = vm.get_cursor_position();
-        assert_eq!(cursor.line, 0);
-        assert_eq!(cursor.column, 5);
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_only_in_insert_mode() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text with empty line
-        vm.insert_text("hello\n\nworld").unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-
-        // Switch to normal mode
-        vm.change_mode(EditorMode::Normal).unwrap();
-
-        // Delete should do nothing in normal mode
-        vm.delete_char_after_cursor().unwrap();
-
-        // Text should be unchanged
-        assert_eq!(vm.get_request_text(), "hello\n\nworld");
-    }
-
-    #[test]
-    fn test_delete_char_after_cursor_only_in_request_pane() {
-        let mut vm = ViewModel::new();
-        vm.change_mode(EditorMode::Insert).unwrap();
-
-        // Insert text in request pane
-        vm.insert_text("hello\n\nworld").unwrap();
-
-        // Set response and switch to response pane
-        vm.set_response(200, "response\n\ntext".to_string());
-        vm.switch_pane(Pane::Response).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 8)).unwrap();
-
-        // Delete should do nothing in response pane
-        vm.delete_char_after_cursor().unwrap();
-
-        // Request text should be unchanged
-        assert_eq!(vm.get_request_text(), "hello\n\nworld");
-        // Response text should also be unchanged (no deletion in response pane)
-        assert_eq!(vm.get_response_text(), "response\n\ntext");
-    }
-
-    // Tests for Visual mode over wrapped lines (Issue #31)
-    #[test]
-    fn test_visual_selection_across_wrapped_lines() {
-        let mut vm = ViewModel::new();
-        // Enable word wrapping with narrow width to force wrapping
-        vm.update_terminal_size(20, 10);
-        vm.set_wrap_enabled(true).unwrap();
-
-        // Insert a long line that will wrap: "This is a very long line that will definitely wrap"
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("This is a very long line that will definitely wrap")
-            .unwrap();
-
-        // Start visual mode at position (0, 5) - at the word "is"
-        vm.change_mode(EditorMode::Normal).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 5)).unwrap();
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        // Move cursor to position (0, 35) - into the wrapped portion
-        vm.set_cursor_position(LogicalPosition::new(0, 35)).unwrap();
-
-        // Verify visual selection spans from position 5 to 35
-        let (start, end, pane) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 5)));
-        assert_eq!(end, Some(LogicalPosition::new(0, 35)));
-        assert_eq!(pane, Some(Pane::Request));
-
-        // Test position selection at various points
-        // Position at column 5 should be selected (start of selection)
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 5), Pane::Request));
-        // Position at column 20 should be selected (middle of selection)
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 20), Pane::Request));
-        // Position at column 35 should be selected (end of selection)
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 35), Pane::Request));
-        // Position at column 4 should not be selected (before start)
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 4), Pane::Request));
-        // Position at column 36 should not be selected (after end)
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 36), Pane::Request));
-    }
-
-    #[test]
-    fn test_visual_selection_in_second_wrapped_line() {
-        let mut vm = ViewModel::new();
-        // Enable word wrapping with narrow width
-        vm.update_terminal_size(20, 10);
-        vm.set_wrap_enabled(true).unwrap();
-
-        // Insert a long line that will wrap into multiple display lines
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("This is the first part and this is the second part that continues")
-            .unwrap();
-
-        // Start visual mode at position (0, 25) - in the second wrapped part
-        vm.change_mode(EditorMode::Normal).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 25)).unwrap();
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        // Extend selection to position (0, 45) - deeper into the wrapped part
-        vm.set_cursor_position(LogicalPosition::new(0, 45)).unwrap();
-
-        // Verify selection is correct
-        let (start, end, _) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 25)));
-        assert_eq!(end, Some(LogicalPosition::new(0, 45)));
-
-        // Test that positions in the wrapped section are correctly identified as selected
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 30), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 40), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 20), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 50), Pane::Request));
-    }
-
-    #[test]
-    #[ignore = "TODO: Fix cursor movement in wrapped lines - skipping from position 13 to 15"]
-    fn test_visual_selection_character_by_character_movement_across_wrapped_lines() {
-        let mut vm = ViewModel::new();
-        // Enable word wrapping with narrow width to force wrapping
-        vm.update_terminal_size(20, 10);
-        vm.set_wrap_enabled(true).unwrap();
-
-        // Insert a long line that will wrap: "This is a very long line"
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("This is a very long line that wraps")
-            .unwrap();
-
-        // Start visual mode at the beginning and test character-by-character movement
-        vm.change_mode(EditorMode::Normal).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 0)).unwrap();
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        // Test moving right character by character through wrapped line segments
-        let start_pos = vm.get_cursor_position();
-
-        // Move cursor right multiple times to cross wrap boundaries
-        for i in 1..=30 {
-            let prev_pos = vm.get_cursor_position();
-            vm.move_cursor_right().unwrap();
-            let current_pos = vm.get_cursor_position();
-
-            // Verify cursor advances properly
-            assert_eq!(
-                current_pos.line, 0,
-                "Cursor should stay on same logical line"
-            );
-
-            // Special case: there's a known issue where cursor jumps from 13 to 15
-            // This happens at position 14 ('y' in "very") in wrapped text
-            // TODO: Investigate and fix this cursor movement issue
-            let expected_column = if i == 14 && prev_pos.column == 13 && current_pos.column == 15 {
-                15 // Accept the jump for now
-            } else if i > 14 && prev_pos.column >= 13 {
-                i + 1 // Adjust for the skip
-            } else {
-                i
-            };
-
-            // After moving right i times from position 0, we should be at expected position
-            assert_eq!(
-                current_pos.column, expected_column,
-                "Cursor should advance by one character each time (iteration {i})"
-            );
-
-            // Verify visual selection is maintained
-            let (visual_start, visual_end, _) = vm.get_visual_selection();
-            assert_eq!(visual_start, Some(start_pos));
-            assert_eq!(visual_end, Some(current_pos));
-        }
-    }
-
-    #[test]
-    fn test_visual_selection_spanning_multiple_logical_lines_with_wrapping() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(15, 10);
-        vm.set_wrap_enabled(true).unwrap();
-
-        // Insert multiple lines where the first line wraps
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("This is a very long first line that wraps\nSecond line")
-            .unwrap();
-
-        // Start visual mode at end of first line (in wrapped portion)
-        vm.change_mode(EditorMode::Normal).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 35)).unwrap();
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        // Extend selection to second line
-        vm.set_cursor_position(LogicalPosition::new(1, 6)).unwrap();
-
-        // Verify selection spans across logical lines
-        let (start, end, _) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 35)));
-        assert_eq!(end, Some(LogicalPosition::new(1, 6)));
-
-        // Test that wrapped portion of first line is selected
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 38), Pane::Request));
-        // Test that beginning of second line is selected
-        assert!(vm.is_position_selected(LogicalPosition::new(1, 3), Pane::Request));
-        // Test boundaries
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 30), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(1, 8), Pane::Request));
-    }
-
-    #[test]
-    fn test_visual_selection_with_word_wrapping_disabled() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(20, 10);
-        vm.set_wrap_enabled(false).unwrap(); // Disable wrapping
-
-        // Insert a long line that would wrap if wrapping was enabled
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("This is a very long line that would wrap but truncates instead")
-            .unwrap();
-
-        // Start visual mode and select part of the long line
-        vm.change_mode(EditorMode::Normal).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 10)).unwrap();
-        vm.change_mode(EditorMode::Visual).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 30)).unwrap();
-
-        // Should work normally without wrapping complications
-        let (start, end, _) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 10)));
-        assert_eq!(end, Some(LogicalPosition::new(0, 30)));
-
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 20), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 5), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 35), Pane::Request));
-    }
-
-    #[test]
-    fn test_visual_selection_reversed_direction_on_wrapped_line() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(20, 10);
-        vm.set_wrap_enabled(true).unwrap();
-
-        // Insert a long line that wraps
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("This is a very long line that will definitely wrap around")
-            .unwrap();
-
-        // Start visual mode at position (0, 30) in wrapped portion
-        vm.change_mode(EditorMode::Normal).unwrap();
-        vm.set_cursor_position(LogicalPosition::new(0, 30)).unwrap();
-        vm.change_mode(EditorMode::Visual).unwrap();
-
-        // Move cursor backwards to position (0, 10) - should handle reversed selection
-        vm.set_cursor_position(LogicalPosition::new(0, 10)).unwrap();
-
-        // Verify selection is normalized (start <= end)
-        let (start, end, _) = vm.get_visual_selection();
-        assert_eq!(start, Some(LogicalPosition::new(0, 30))); // Original start position
-        assert_eq!(end, Some(LogicalPosition::new(0, 10))); // Current cursor position
-
-        // Test that positions between 10 and 30 are selected (inclusive)
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 15), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 25), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 10), Pane::Request));
-        assert!(vm.is_position_selected(LogicalPosition::new(0, 30), Pane::Request));
-        // Test boundaries
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 8), Pane::Request));
-        assert!(!vm.is_position_selected(LogicalPosition::new(0, 32), Pane::Request));
-    }
-
-    // Tests for PaneState word navigation methods
-    #[test]
-    fn test_pane_state_find_next_word_position_should_find_next_word_on_same_line() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with multiple words
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world test").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From position (0, 0), should find "world" at (0, 6)
-        let result = pane_state.find_next_word_position((0, 0));
-        assert_eq!(result, Some((0, 6)));
-
-        // From position (0, 6), should find "test" at (0, 12)
-        let result = pane_state.find_next_word_position((0, 6));
-        assert_eq!(result, Some((0, 12)));
-    }
-
-    #[test]
-    fn test_pane_state_find_next_word_position_should_find_word_on_next_line() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with multiple lines
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello\n  world test").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From end of first line, should find "world" on next line
-        let result = pane_state.find_next_word_position((0, 5));
-        assert_eq!(result, Some((1, 2))); // "world" starts at column 2 (after spaces)
-    }
-
-    #[test]
-    fn test_pane_state_find_next_word_position_should_return_none_at_end() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert single word
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From end of text, should return None
-        let result = pane_state.find_next_word_position((0, 5));
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_pane_state_find_previous_word_position_should_find_previous_word_on_same_line() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with multiple words
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world test").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From position (0, 12) "test", should find "world" at (0, 6)
-        let result = pane_state.find_previous_word_position((0, 12));
-        assert_eq!(result, Some((0, 6)));
-
-        // From position (0, 6) "world", should find "hello" at (0, 0)
-        let result = pane_state.find_previous_word_position((0, 6));
-        assert_eq!(result, Some((0, 0)));
-    }
-
-    #[test]
-    fn test_pane_state_find_previous_word_position_should_find_word_on_previous_line() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with multiple lines
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world\ntest").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From beginning of second line, should find "world" on previous line
-        let result = pane_state.find_previous_word_position((1, 0));
-        assert_eq!(result, Some((0, 6))); // "world" starts at column 6
-    }
-
-    #[test]
-    fn test_pane_state_find_previous_word_position_should_return_none_at_beginning() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert single word
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From beginning of text, should return None
-        let result = pane_state.find_previous_word_position((0, 0));
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_pane_state_find_end_of_word_position_should_find_end_of_current_word() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with multiple words
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world test").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From position (0, 0) in "hello", should find end at (0, 4)
-        let result = pane_state.find_end_of_word_position((0, 0));
-        assert_eq!(result, Some((0, 4))); // 'o' in "hello"
-
-        // From position (0, 6) in "world", should find end at (0, 10)
-        let result = pane_state.find_end_of_word_position((0, 6));
-        assert_eq!(result, Some((0, 10))); // 'd' in "world"
-    }
-
-    #[test]
-    fn test_pane_state_find_end_of_word_position_should_find_next_word_end_when_at_word_end() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with multiple words
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello world test").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From end of "hello" (position 4), should find end of "world" at (0, 10)
-        let result = pane_state.find_end_of_word_position((0, 4));
-        assert_eq!(result, Some((0, 10))); // 'd' in "world"
-    }
-
-    #[test]
-    fn test_pane_state_find_end_of_word_position_should_return_none_at_end() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert single word
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // From end of text, should return None
-        let result = pane_state.find_end_of_word_position((0, 4));
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_pane_state_word_navigation_should_handle_punctuation() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with punctuation
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello, world! test").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // Debug: let's see what the actual text looks like
-        // "hello, world! test"
-        //  01234567890123456789
-        //  h=0, ,=5, w=7, !=12, t=14
-
-        // Test next word navigation with punctuation
-        let result = pane_state.find_next_word_position((0, 0)); // from "hello"
-                                                                 // Current algorithm behavior: finds the comma at position 5
-                                                                 // This matches vim's behavior where punctuation is treated as a separate "word"
-        assert_eq!(result, Some((0, 5))); // finds comma
-
-        // From comma, should find "world"
-        let result = pane_state.find_next_word_position((0, 5)); // from comma
-        assert_eq!(result, Some((0, 7))); // should find "world"
-
-        let result = pane_state.find_next_word_position((0, 7)); // from "world"
-        assert_eq!(result, Some((0, 12))); // finds exclamation at position 12
-
-        // Test previous word navigation with punctuation
-        let result = pane_state.find_previous_word_position((0, 14)); // from "test"
-        assert_eq!(result, Some((0, 7))); // should find "world"
-
-        let result = pane_state.find_previous_word_position((0, 7)); // from "world"
-        assert_eq!(result, Some((0, 0))); // should find "hello"
-    }
-
-    #[test]
-    fn test_pane_state_word_navigation_should_handle_empty_lines() {
-        let mut vm = ViewModel::new();
-        vm.update_terminal_size(80, 24);
-        vm.set_wrap_enabled(false).unwrap();
-
-        // Insert text with empty lines
-        vm.change_mode(EditorMode::Insert).unwrap();
-        vm.insert_text("hello\n\nworld").unwrap();
-
-        let pane_state = &vm.panes[Pane::Request];
-
-        // Should skip empty line and find "world"
-        let result = pane_state.find_next_word_position((0, 5)); // from end of "hello"
-        assert_eq!(result, Some((2, 0))); // "world" on line 2
-
-        // Should skip empty line going backwards
-        let result = pane_state.find_previous_word_position((2, 0)); // from "world"
-        assert_eq!(result, Some((0, 0))); // "hello" on line 0
     }
 }
